% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abmi.camera.extras.R
\docType{package}
\name{abmi.camera.extras}
\alias{abmi.camera.extras}
\alias{abmi.camera.extras-package}
\title{abmi.camera.extras: Functions and data to make use of ABMI camera deployment data}
\description{
This packages provides some functionality to interact with and explore ABMI camera data.
It includes some sample data, important lookup tables, and a few functions to serve basic needs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mabecker89/abmi.camera.extras}
  \item Report bugs at \url{https://github.com/mabecker89/abmi.camera.extras/issues}
}

}
\author{
Marcus Becker \email{mabecker@ualberta.ca}
}
