% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_loc.R
\docType{data}
\name{abmi_deployment_locations}
\alias{abmi_deployment_locations}
\title{ABMI camera deployments public locations}
\format{
A data.frame with 5 columns
\describe{
\item{Deployment}{The camera deployment ID. Format is <Project Code>-<Site>-<Station>, e.g. ABMI-1-NE.}
\item{NearestSite}{The corresponding nearest ABMI sampling site.}
\item{Year}{The year the camera was deployed.}
\item{Public_Lat}{The public latitude}
\item{Public_Long}{The public longitude}
}
}
\source{
Alberta Biodiversity Monitoring Institute
}
\usage{
abmi_deployment_locations
}
\description{
This dataset provides the lat/long coordinates associated with the public
locations of ABMI camera deployments.
}
\seealso{
\url{https://www.abmi.ca/home/data-analytics/DA-top-about-data/About-ABMI-Data/Survey-Site-Locations.html}
}
\keyword{data}
