% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace_join_dens.R
\name{ace_join_dens}
\alias{ace_join_dens}
\title{Join estimated ABMI single-deployment density values}
\usage{
ace_join_dens(x, species, samp_per, nest = FALSE)
}
\arguments{
\item{x}{a dataframe of deployments of interest, including at minimum the deployment id and year}

\item{species}{character; vector of species of interest (common name). If left blank, all species available will be returned.}

\item{samp_per}{character; vector of sampling period of interest. Available periods include the years 2013-2019, inclusive. If left blank, estimates from all available sampling periods will be returned.}

\item{nest}{logical; if TRUE, a dataframe nested by species (common name) is returned.}
}
\value{
Tidy dataframe of deployments in sampling period(s) specified with two appended columns: species and estimated density.
}
\description{
Join estimated ABMI single-deployment density values
}
\details{
valid values for argument \code{species} currently are:
\itemize{
 \item White-tailed Deer
 \item Mule deer
 \item Moose
 \item Elk (wapiti)
 \item Black Bear
 \item Coyote
 \item Pronghorn
 \item Snowshoe Hare
 \item Woodland Caribou
 \item Canada Lynx
 \item Gray Wolf
 }

valid values for argument \code{samp_per} currently are:
\itemize{
 \item 2013
 \item 2014
 \item 2015
 \item 2016
 \item 2017
 \item 2018
 \item 2019
 }
}
\examples{
library(dplyr)
# Dataframe of deployment names and year
df <- data.frame(name = c("ABMI-633-NE",
                          "ABMI-633-NW",
                          "ABMI-633-SE",
                          "ABMI-633-SW"),
                 samp_per = c(2015, 2015, 2015, 2015)) \%>\%
                 mutate_if(is.factor, as.numeric)
# Join density estimates (e.g. Moose in 2015)
df_densities <- ace_join_dens(df, species = "Moose", samp_per = "2015", nest = FALSE)
}
