% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cam.R
\name{get_cam_within_aoi}
\alias{get_cam_within_aoi}
\title{Subset ABMI camera deployments by a specified spatial area of interest}
\usage{
get_cam_within_aoi(x, cols = NULL, keep.all = FALSE)
}
\arguments{
\item{x}{The area of interest as an sf, sfc, or sp (SpatialPolygonsDataFrame) object}

\item{cols}{character; indicates which columns (attributes) from x to be appended to the deployments dataframe}

\item{keep.all}{logical; if TRUE all deployments are kept, otherwise only those within x are kept}
}
\value{
a dataframe of deployments within the supplied area of interest
}
\description{
Subset ABMI camera deployments by a specified spatial area of interest
}
\examples{
# Example aoi of four Wildlife Management Units (WMUs) in Alberta:
wmu_sample <- st_read(system.file("extdata/wmu_sample.shp", package = "abmi.camera.extras"))
# Obtain ABMI deployments in sample WMUs, keeping unit name
wmu_sample_deployments <- subset_deployments_by_aoi(wmu_sample, cols = "WMUNIT_NAM")
# Plot results
wmu_sample <- st_transform(wmu_sample, "+init=epsg:4326")
plot(wmu_sample_deployments$geometry, pch = 21, cex = 0.7, col = "blue", bg = "gray80")
plot(wmu_sample$geometry, border = "gray20", col = NA, add = TRUE)
}
\author{
Marcus Becker
}
