% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{simul_ci}
\alias{simul_ci}
\title{Utility functions}
\usage{
simul_ci(prob, trials, adj, agp, agp.se)
}
\arguments{
\item{prob}{the occupancy rate}

\item{trials}{the number of deployments}

\item{adj}{the number of occupied sites}

\item{agp}{abundance given presence}

\item{agp.se}{the standard error of the abundance given presence}
}
\value{
numeric of type double with length of 10000 (the number of simulations)
}
\description{
Utility functions
}
\author{
Marcus Becker
}
