% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{species}
\alias{species}
\title{Mammal species included in this package.}
\format{
A character vector
}
\source{
Alberta Biodiversity Monitoring Institute
}
\usage{
species
}
\description{
Eleven (11) common mammal species for which density estimates are provided as part of this package
}
\seealso{
\url{https://github.com/ABbiodiversity/mammals-camera}
}
\keyword{data}
