% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace_get_cam.R
\name{ace_get_cam}
\alias{ace_get_cam}
\title{Subset camera deployments by a specified spatial area of interest}
\usage{
ace_get_cam(aoi, group_id = NULL, dep = NULL, coords = NULL, crs = 4326)
}
\arguments{
\item{aoi}{The area of interest as an sf, sfc, or sp (SpatialPolygonsDataFrame) object}

\item{group_id}{If aoi contains multiple polygons, name of the attribute to be appended to the output dataframe as identifier}

\item{dep}{A dataframe of camera deployment locations as coordinate points; defaults to NULL, in which case ABMI camera deployment locations are used.}

\item{coords}{If dep is specified, names of the of numeric columns holding coordinates}

\item{crs}{coordinate reference system; integer with the EPSG code, or character with proj4string; defaults to 4326.}
}
\value{
A dataframe of camera deployments within the supplied area of interest (aoi)
}
\description{
Subset camera deployments by a specified spatial area of interest
}
\examples{

library(sf)
# Example aoi of four Wildlife Management Units (WMUs) in Alberta:
wmu_sample <- st_read(system.file("extdata/wmu_sample.shp", package = "abmi.camera.extras"))
# Obtain ABMI deployments in sample WMUs, keeping unit name
wmu_sample_deployments <- ace_get_cam(wmu_sample, group_id = WMUNIT_NAM)
# Plot results
wmu_sample <- st_transform(wmu_sample, "+init=epsg:4326")
plot(wmu_sample_deployments$geometry, pch = 21, cex = 0.7, col = "blue", bg = "gray80")
plot(wmu_sample$geometry, border = "gray20", col = NA, add = TRUE)

}
