% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace_join_dens.R
\name{ace_join_dens}
\alias{ace_join_dens}
\title{Join estimated single-deployment density values}
\usage{
ace_join_dens(x, species, year, nest = FALSE)
}
\arguments{
\item{x}{a dataframe of deployments of interest, including at minimum the deployment id and year}

\item{species}{character; vector of species of interest (common name). If left blank, all species available will be returned.}

\item{year}{character; vector of years of interest (2013-2018). If left blank, all years available will be returned.}

\item{nest}{logical; if TRUE, a dataframe nested by species (common name) is returned.}
}
\value{
Tidy dataframe of deployments in year(s) specified with two appended columns: species and estimated density.
}
\description{
Join estimated single-deployment density values
}
\details{
valid values for argument \code{species} currently are:
\itemize{
 \item White-tailed Deer
 \item Mule deer
 \item Moose
 \item Elk (wapiti)
 \item Black Bear
 \item Coyote
 \item Pronghorn
 \item Snowshoe Hare
 \item Woodland Caribou
 \item Canada Lynx
 \item Gray Wolf
 }
}
\examples{
library(dplyr)
# Dataframe of deployments and year
df <- data.frame(Deployment = c("ABMI-633-NE",
                                "ABMI-633-NW",
                                "ABMI-633-SE",
                                "ABMI-633-SW"),
                 Year = c(2015, 2015, 2015, 2015)) \%>\%
                 mutate_if(is.factor, as.character)
# Join density estimates (e.g. Moose in 2015)
df_densities <- ace_join_dens(df, species = "Moose", year = "2015", nest = FALSE)
}
\author{
Marcus Becker
}
